qca-nss-drv: Fix wifi offload regression
The option for wifi offload is `NSS_DRV_WIFIOFFLOAD_ENABLE` but the
package Makefile was showing it as `NSS_DRV_WIFI_ENABLE` but then enabling the
flag as `NSS_DRV_WIFI_ENABLE`. Incorrectly assumed `NSS_DRV_WIFI_ENABLE`
was the global option. Reverted the change and updated the option to use
`NSS_DRV_WIFIOFFLOAD_ENABLE`.