nss-drv: 11.4 split out formatting from 12.4 backbort
This is mostly a cosmetic change and meant to make it easier to see
the to see history of commits.

The currently published NSS FW on QUIC's repo is:

NSS.HK.11.4.0.5-5-R which is on branch caf_migration/NHSS.QSDK.11.4.0.5

However, it's not the latest of the `11.4.0.5` release which is

NSS.HK.11.4.0.5-6-R which is on branch `origin/NSS_11.4.0.5`

Although the versions look to be minor releases. NSS driver shows
significant changes in the headers files concerning `wifi` and `wifi
mesh`.

When doing a diff between the 2 releases, it was hard to tell exactly
what changed because most of it seemed to be non-functional changes,
formatting, etc. This patch is only meant to make the files visually
similiar between the two releases, to better show functional changes.

Signed-off-by: Sean Khan <datapronix@protonmail.com>